# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API d'un service d'authentification",
        "version": "1.0"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enregistrement d’un nouvel utilisateur",
                "description": "La route est throttled à 5 requête par minute.",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "login",
                                    "password",
                                    "phone"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Bob"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Marley"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "test123456"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "418-555-1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "first_name": "Jeremie",
                                        "last_name": "Paquin",
                                        "email": "test@example.ca",
                                        "login": "test",
                                        "phone": "418-555-1234"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The email field must be a valid email address.",
                                    "errors": {
                                        "email": [
                                            "The email field must be a valid email address."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authentification d’un utilisateur existant",
                "description": "La route est throttled à 5 requête par minute.",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "test123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "token": "..."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The password field is required.",
                                    "errors": {
                                        "password": [
                                            "The password field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authorisé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Révocation des tokens de l’utilisateur",
                "description": "La route est throttled à 5 requête par minute.",
                "operationId": "b46f8611fc7d1123f2a3d573cee661c4",
                "responses": {
                    "default": {
                        "description": "Default JSON response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Utilisateur deconnecté"
                    },
                    "401": {
                        "description": "Non authorisé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rafraichissement du token existant",
                "description": "La route est throttled à 5 requête par minute.",
                "operationId": "2c27d311a2876560fcb5a754b556cdc5",
                "responses": {
                    "200": {
                        "description": "Token rafraîchi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "token": "..."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authorisé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Accès à l’utilisateur connecté",
                "description": "La route est throttled à 5 requête par minute.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Information de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "first_name": "Jeremie",
                                        "last_name": "Paquin",
                                        "email": "test@example.ca",
                                        "login": "test",
                                        "phone": "418-555-1234"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authorisé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Entrer le token sous le format (Bearer TOKEN)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}